+++ src/main/scala/ciir/ts/SimilarTerms.scala
@@ -5,13 +5,13 @@ import gnu.trove.set.hash.{TIntHashSet, TLongHashSet}
 import collection.mutable.ArrayBuilder
 
 object Statistics {
-  def mean(xs: Array[Long]) = {
+  def mean(xs: Array[Int]) = {
     xs.sum.toDouble / xs.size.toDouble
   }
   def mean(xs: Array[Double]) = {
     xs.sum / xs.size
   }
-  def median(xs: Array[Long]): Double = {
+  def median(xs: Array[Int]): Double = {
     val tmp = xs.sorted
     val mid = tmp.size/2
     if(tmp.size % 2 == 1) {
@@ -20,21 +20,21 @@ object Statistics {
       (tmp(mid-1) + tmp(mid)) / 2
     }
   }
-  def variance(xs: Array[Long]) = {
+  def variance(xs: Array[Int]) = {
     val mu = mean(xs)
     mean(xs.map(x =>{
       val off = x.toDouble - mu
       off*off
     }))
   }
-  def maxIndex(xs: Array[Long]) = { 
+  def maxIndex(xs: Array[Int]) = { 
     xs.zipWithIndex.maxBy(_._1)._2
   }
-  def minIndex(xs: Array[Long]) = { 
+  def minIndex(xs: Array[Int]) = { 
     xs.zipWithIndex.minBy(_._1)._2
   }
   
-  def summary(data: Array[Long], base: Int=0) = {
+  def summary(data: Array[Int], base: Int=0) = {
     println("  total: "+data.sum)
     println("  max: "+data.max+" in "+(base+maxIndex(data)))
     println("  mean: "+mean(data))
